Cleaned up serviceEventsForm to pass the database function down
